refactor: fix using theme inside emotion css
